+++ examples/client/processor/location_information.cpp
@@ -1,6 +1,9 @@
 #include "location_information.hpp"
 #include "client.hpp"
 
+#include <sstream>
+
+#include <time/utc.hpp>
 #include <loglet/loglet.hpp>
 
 LOGLET_MODULE2(p, li);
@@ -14,4 +17,69 @@ void LocationCollector::inspect(streamline::System&, DataType const& location) N
 void MetricsCollector::inspect(streamline::System&, DataType const& metrics) NOEXCEPT {
     VSCOPE_FUNCTION();
     mProgram.update_gnss_metrics(metrics);
+}
+
+static void horizontal_accuracy_json(std::stringstream& result, lpp::HorizontalAccuracy const& horizontal_accuracy) {
+    result << "\"semi-major\": " << horizontal_accuracy.semi_major;
+    result << ",\"semi-minor\": " << horizontal_accuracy.semi_minor;
+    result << ",\"orientation\": " << horizontal_accuracy.orientation;
+    result << ",\"confidence\": " << horizontal_accuracy.confidence;
+}
+
+static void vertical_accuracy_json(std::stringstream& result, lpp::VerticalAccuracy const& vertical_accuracy) {
+    result << "\"uncertainty\": " << vertical_accuracy.uncertainty;
+    result << ",\"confidence\": " << vertical_accuracy.confidence;
+}
+
+void LocationOutput::inspect(streamline::System&, DataType const& location) NOEXCEPT {
+    VSCOPE_FUNCTION();
+    
+    std::stringstream result;
+    result << "{";
+    result << "\"type\": \"location\",";
+    result << ",\"time\": \"" << ts::Utc{location.time}.rfc3339() << "\"";
+    if(location.location.has_value()) {
+        const auto& shape = location.location.const_value();
+        result << ",\"location\": {";
+        switch(shape.kind) {
+            case lpp::LocationShape::Kind::HighAccuracyEllipsoidPointWithUncertaintyEllipse:
+                result << "\"type\": \"high-accuracy-ellipsoid-point-with-uncertainty-ellipse\"";
+                result << ",\"latitude\": " << shape.data.haepue.latitude;
+                result << ",\"longitude\": " << shape.data.haepue.longitude;
+                result << ",\"horizontal-accuracy\": ";
+                horizontal_accuracy_json(result, shape.data.haepue.horizontal_accuracy);
+                break;
+            case lpp::LocationShape::Kind::HighAccuracyEllipsoidPointWithAltitudeAndUncertaintyEllipsoid:
+                result << "\"type\": \"high-accuracy-ellipsoid-point-with-altitude-and-uncertainty-ellipsoid\"";
+                result << ",\"latitude\": " << shape.data.haepaue.latitude;
+                result << ",\"longitude\": " << shape.data.haepaue.longitude;
+                result << ",\"altitude\": " << shape.data.haepaue.altitude;
+                result << ",\"horizontal-accuracy\": ";
+                horizontal_accuracy_json(result, shape.data.haepaue.horizontal_accuracy);
+                result << ",\"vertical-accuracy\": ";
+                vertical_accuracy_json(result, shape.data.haepaue.vertical_accuracy);
+                break;
+            default:
+                result << "\"type\": \"unknown\"";
+                break;
+        }
+
+        result << "}";
+    }
+    if(location.velocity.has_value()) {
+        result << ",\"velocity\": {";
+        result << "}";
+    }
+    result << "}\n";
+    auto sentence = result.str();
+    auto data     = reinterpret_cast<uint8_t const*>(sentence.data());
+    auto size     = sentence.size();
+    for (auto const& output : mOutput.outputs) {
+        if (!output.location_support()) continue;
+        if (output.print) {
+            XINFOF(OUTPUT_PRINT_MODULE, "location: %zd bytes", size);
+        }
+
+        output.interface->write(data, size);
+    }
 }